+++ plot/monitor_test.go
@@ -50,4 +50,5 @@ func ExampleNewMonitorWindow() {
 	// Uncomment the next line. It is commented out as the CI has no display device to test the model run.
 
 	// pixelgl.Run(m.Run)
+	// Output:
 }