+++ plot/time_series.go
@@ -7,7 +7,6 @@ import (
 	"github.com/faiface/pixel/pixelgl"
 	"github.com/mlange-42/arche-model/observer"
 	"github.com/mlange-42/arche/ecs"
-	"golang.org/x/image/colornames"
 	"gonum.org/v1/plot"
 	"gonum.org/v1/plot/plotter"
 	"gonum.org/v1/plot/vg"
@@ -15,19 +14,13 @@ import (
 	"gonum.org/v1/plot/vg/vgimg"
 )
 
-var defaultColors = []color.Color{
-	colornames.Blue,
-	colornames.Orange,
-	colornames.Green,
-	colornames.Purple,
-	colornames.Red,
-	colornames.Turquoise,
-}
-
 // TimeSeries plot drawer.
+//
+// Creates a line series per column of the observer.
+// Adds one row to the data per update.
 type TimeSeries struct {
 	Observer       observer.Row // Observer providing a data row per update.
-	UpdateInterval int          // Interval for updating the observer, in model ticks. Optional.
+	UpdateInterval int          // Interval for getting data from the the observer, in model ticks. Optional.
 
 	headers []string
 	series  []plotter.XYs
@@ -35,6 +28,7 @@ type TimeSeries struct {
 	step    int64
 }
 
+// append a y value to each series, with a common x value.
 func (t *TimeSeries) append(x float64, values []float64) {
 	for i := 0; i < len(t.series); i++ {
 		t.series[i] = append(t.series[i], plotter.XY{X: x, Y: values[i]})
@@ -54,8 +48,8 @@ func (t *TimeSeries) Initialize(w *ecs.World, win *pixelgl.Window) {
 
 // Update the drawer.
 func (t *TimeSeries) Update(w *ecs.World) {
+	t.Observer.Update(w)
 	if t.UpdateInterval <= 1 || t.step%int64(t.UpdateInterval) == 0 {
-		t.Observer.Update(w)
 		t.append(float64(t.step), t.Observer.Values(w))
 	}
 	t.step++